pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "frolkin28", email = "frolkin2801@gmail.com"},
]
dependencies = [
    "flask>=3.0.2",
    "Flask-SQLAlchemy>=3.1.1",
    "psycopg2-binary>=2.9.9",
    "Flask-Migrate>=4.0.5",
    "pyyaml>=6.0.1",
    "marshmallow>=3.21.0",
    "apispec>=6.5.0",
    "Flask-JWT-Extended>=4.6.0",
    "apispec-webframeworks>=1.0.0",
    "flask-swagger-ui>=4.11.1",
    "gunicorn>=21.2.0",
    "eventlet>=0.35.2",
    "flask-cors>=4.0.0",
    "python-dotenv>=1.0.1",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

